move independent files
