--- saxs/algo/README.md
@@ -1,16 +1,0 @@
-- deleting in preallocated arr
-- abstr 
-- custom phase
-
-
-phase proccessing
-```python
-phases = Fastdw(filename, current_session, defined_phases, class_names, data[peaks.file])
-
-phases.preset_plot()
-
-phases.data_preparing()
-
-phases.alignement()
-
-data[peaks.file] = phases.gathering()
