--- saxs/algo/data/stage_objects.py
@@ -1,17 +1,0 @@
-#
-# Created by Isai GORDEEV on 20/09/2025.
-#
-
-
-from dataclasses import dataclass, field
-from typing import Any, Dict
-
-from saxs.algo.data.abstract_data import AData
-
-
-@dataclass(frozen=True)
-class AbstractStageMetadata(AData):
-    data: Dict[str, Any] = field(default_factory=dict)
-
-    def describe(self) -> str:
-        return f"Stage metadata with keys: {list(self.data.keys())}"
