--- saxs/algo/settings_processing.py
@@ -1,18 +1,0 @@
-# to be implemented unique for each test_processing_data + manual adjusting
-START = 0.02
-BACKGROUND_COEF = 0.7  # normal 0.7
-SIGMA_FILTER = 1.5
-SIGMA_FITTING = 0.3
-TRUNCATE = 4
-PROMINENCE = 0.6
-INFINITY = 100
-WINDOWSIZE = 6
-RESOLUTION_FACTOR = 1.4
-
-FILENAME = "075966_treated_xye"  # pointwise classification
-EXTENSION = ".csv"
-ANALYSE_DIR = "results"
-ANALYSE_DIR_SESSIONS = "plots/"
-ANALYSE_DIR_SESSIONS_RESULTS = "session_results/"
-PHASES_DIR = "phases.json"
-DATA_DIR = "data_processing_test/"
